Move batch validation to BatchValidator module from BatchesController
